feature/subshell funcionando na execução (signal em comando e heredoc também). Chamar parenteses vazios causa erro no fechamento do processo filho, pois não tem o que executar e buga, por isso, tem que fazer dar erro no parsing. Tanto para isso, quanto para pipe/heredoc/redirect depois ou antes dos parenteses

+++ Signal/signal.c
@@ -47,6 +47,12 @@ int	handle_sigint_in_fork(int status, pid_t pid)
 		if (WTERMSIG(status) == SIGINT)
 		{
 			get_data()->exited_in_fork = true;
+			if (get_data()->is_subshell)
+			{
+				free_tree(get_data()->head);
+				free_list(get_data()->env);
+				exit(130);
+			}
 			kill(SIGINT, pid);
 			return (1);
 		}
